move avg and inttime to where they belong
